+++ src/meta.rs
@@ -6,20 +6,20 @@ use crate::utils::now;
 use crate::Error;
 
 /// Holds information about a cached resource.
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub struct Meta {
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub(crate) struct Meta {
     /// The original resource name.
-    pub resource: String,
+    pub(crate) resource: String,
     /// Path to the cached resource.
-    pub resource_path: PathBuf,
+    pub(crate) resource_path: PathBuf,
     /// Path to the serialized meta.
-    pub meta_path: PathBuf,
+    pub(crate) meta_path: PathBuf,
     /// The ETAG of the resource from the time it was cached, if there was one.
-    pub etag: Option<String>,
+    pub(crate) etag: Option<String>,
     /// Time that the freshness of this cached resource will expire.
-    pub expires: Option<f64>,
+    pub(crate) expires: Option<f64>,
     /// Time this version of the resource was cached.
-    pub creation_time: f64,
+    pub(crate) creation_time: f64,
 }
 
 impl Meta {
@@ -53,40 +53,40 @@ impl Meta {
         meta_path
     }
 
+    pub(crate) fn get_extraction_path(&self) -> PathBuf {
+        let dirname = format!(
+            "{}-extracted",
+            self.resource_path.file_name().unwrap().to_str().unwrap()
+        );
+        self.resource_path.parent().unwrap().join(dirname)
+    }
+
     pub(crate) fn to_file(&self) -> Result<(), Error> {
         let serialized = serde_json::to_string(self).unwrap();
         fs::write(&self.meta_path, &serialized[..])?;
         Ok(())
     }
 
-    /// Get the `Meta` from a cached resource.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use cached_path::{cached_path, Meta};
-    ///
-    /// let resource = "https://github.com/epwalsh/rust-cached-path/blob/master/README.md";
-    /// let path = cached_path(resource).unwrap();
-    /// let meta = Meta::from_cache(&path).unwrap();
-    /// assert_eq!(&meta.resource[..], resource);
-    /// ```
-    pub fn from_cache(resource_path: &Path) -> Result<Self, Error> {
+    pub(crate) fn from_cache(resource_path: &Path) -> Result<Self, Error> {
         let meta_path = Meta::meta_path(resource_path);
         Meta::from_path(&meta_path)
     }
 
     /// Read `Meta` from a path.
     pub(crate) fn from_path(path: &Path) -> Result<Self, Error> {
+        if !path.is_file() {
+            return Err(Error::CacheCorrupted(format!("missing meta at {:?}", path)));
+        }
         let serialized = fs::read_to_string(path)?;
-        let meta: Meta = serde_json::from_str(&serialized[..]).unwrap();
+        let meta: Meta = serde_json::from_str(&serialized[..])
+            .map_err(|e| Error::CacheCorrupted(format!("invalid meta at {:?}: {:?}", path, e)))?;
         Ok(meta)
     }
 
     /// Check if resource is still fresh. Passing a `Some` value for
     /// `freshness_lifetime` will override the expiration time (if there is one)
     /// of this resource.
-    pub fn is_fresh(&self, freshness_lifetime: Option<u64>) -> bool {
+    pub(crate) fn is_fresh(&self, freshness_lifetime: Option<u64>) -> bool {
         if let Some(lifetime) = freshness_lifetime {
             let expiration_time = self.creation_time + (lifetime as f64);
             expiration_time > now()